BAEL-9260: Hashmap Implementation to Count the Occurrences of Each Character in Java

+++ core-java-modules/core-java-string-algorithms-5/src/test/java/com/baeldung/countingchars/CharacterFrequencyCounterUnitTest.java
@@ -22,12 +22,13 @@ public class CharacterFrequencyCounterUnitTest {
     @Test
     public void givenSimpleInput_whenCountingCharactersWithStreams_thenReturnsCorrectFrequencies() {
         String input = "test";
-        Map<Character, Long> result = CharacterFrequencyCounter.countCharactersWithStreams(input);
+        Map<Character, Integer> result = CharacterFrequencyCounter.countCharactersWithStreams(input);
 
-        assertEquals(Long.valueOf(2), result.get('t'));
-        assertEquals(Long.valueOf(1), result.get('e'));
-        assertEquals(Long.valueOf(1), result.get('s'));
+        assertEquals(Integer.valueOf(2), result.get('t'));
+        assertEquals(Integer.valueOf(1), result.get('e'));
+        assertEquals(Integer.valueOf(1), result.get('s'));
         assertEquals(3, result.size());
     }
+
 
 }
